Create Linear_and_Knn15_model.R
R code to implement k-NN and linear regression algorithms, and compare the classification performance of the two algorithms on the data you generated for below problem. Show the decision boundary and misclassification error from each
classifier. Use k=15 for k-NN.
Suppose there are two classes of data distributed on a 2-dimensional space. Class one has normal distribution N(1; ), and class two has normal distribution N(2; ), where  is the 2  2 identity matrix, and 1; 2 2 R2. The two classes have the same covariance matrix but different means.
Assume the generative densities are unknown, but you are given 100 data points for each class. Use the generative densities as 1 = (1; 0), and 2 = (0; 1) to generate the data, and then draw the decision boundary based on data.