Add files via upload
Computer program to perform a local linear discriminant analysis. At each query point x0, the training data receive weights K(x0; xi) from a weighting kernel, and the ingredients for the linear discriminant function are computed by weighted averages. Try out your program on the zipcode data, and show the training and test errors for a series of five pre-chosen values of . The zipcode data are provided. Due to the large size of the dataset, you are only considering the classification of three digits, 0, 1, and 2. Use a 70-30 split for training and testing data.
Hint: to localize the function (4.10), the three ingredients